folder restructure and using ctrls and routes complete
