+++ src/routes.js
@@ -3,7 +3,6 @@ const { usersController, chatCmdsController, socketController } = require('./con
 
 routes.put('/login', usersController.login)
 routes.get('/onlineUsers', usersController.get)
-routes.get('/getCmds', chatCmdsController.get)
 routes.put('/addCmd', chatCmdsController.put)
 routes.get('/socketCmds', socketController.get)
